Struct ptp::PtpCamera

pub struct PtpCamera<'a> { /* fields omitted */ }

Methods

impl<'a> PtpCamera<'a>

pub fn new(device: &Device<'a>) -> Result<PtpCamera<'a>, Error>

pub fn command(
&mut self,
code: CommandCode,
params: &[u32],
data: Option<&[u8]>,
timeout: Option<Duration>
) -> Result<Vec<u8>, Error>

execute a PTP transaction. consists of the following phases:

command

command data (optional, if data is Some)

response data (optional, if response contains a payload)

response status NB: each phase involves a separate USB transfer, and timeout is used for each phase, so the total time taken may be greater than timeout.

pub fn get_objectinfo(
&mut self,
handle: u32,
timeout: Option<Duration>
) -> Result<PtpObjectInfo, Error>

pub fn get_object(
&mut self,
handle: u32,
timeout: Option<Duration>
) -> Result<Vec<u8>, Error>

pub fn get_partialobject(
&mut self,
handle: u32,
offset: u32,
max: u32,
timeout: Option<Duration>
) -> Result<Vec<u8>, Error>

pub fn delete_object(
&mut self,
handle: u32,
timeout: Option<Duration>
) -> Result<(), Error>

pub fn power_down(&mut self, timeout: Option<Duration>) -> Result<(), Error>

pub fn get_objecthandles(
&mut self,
storage_id: u32,
handle_id: u32,
filter: Option<u32>,
timeout: Option<Duration>
) -> Result<Vec<u32>, Error>

pub fn get_objecthandles_root(
&mut self,
storage_id: u32,
filter: Option<u32>,
timeout: Option<Duration>
) -> Result<Vec<u32>, Error>

pub fn get_objecthandles_all(
&mut self,
storage_id: u32,
filter: Option<u32>,
timeout: Option<Duration>
) -> Result<Vec<u32>, Error>

pub fn get_numobjects(
&mut self,
storage_id: u32,
handle_id: u32,
filter: Option<u32>,
timeout: Option<Duration>
) -> Result<u32, Error>

pub fn get_storage_info(
&mut self,
storage_id: u32,
timeout: Option<Duration>
) -> Result<PtpStorageInfo, Error>

pub fn get_storageids(
&mut self,
timeout: Option<Duration>
) -> Result<Vec<u32>, Error>

pub fn get_numobjects_roots(
&mut self,
storage_id: u32,
filter: Option<u32>,
timeout: Option<Duration>
) -> Result<u32, Error>

pub fn get_numobjects_all(
&mut self,
storage_id: u32,
filter: Option<u32>,
timeout: Option<Duration>
) -> Result<u32, Error>

pub fn get_device_info(
&mut self,
timeout: Option<Duration>
) -> Result<PtpDeviceInfo, Error>

pub fn open_session(&mut self, timeout: Option<Duration>) -> Result<(), Error>

pub fn close_session(&mut self, timeout: Option<Duration>) -> Result<(), Error>

pub fn disconnect(&mut self, timeout: Option<Duration>) -> Result<(), Error>